KS-2-5 added to.Double to percent calculation

+++ src/main/kotlin/Lesson2/lesson2_task5.kt
@@ -9,7 +9,7 @@ const val PERCENT_FORMATTER = 100
 
 fun main() {
 
-    val interestRate: Double = INTEREST_RATE_PERCENTAGE / PERCENT_FORMATTER
+    val interestRate: Double = INTEREST_RATE_PERCENTAGE.toDouble() / PERCENT_FORMATTER
     val initialSum = 70000
     val years = 20
